implement exchanging

+++ app/src/main/java/dev/sdex/currencyexchanger/domain/model/ExchangeRate.kt
@@ -1,6 +1,8 @@
 package dev.sdex.currencyexchanger.domain.model
 
+import java.math.BigDecimal
+
 data class ExchangeRate(
     val currency: String,
-    val rate: Double,
+    val rate: BigDecimal,
 )
